requirements.txt
pyyaml
googlesearch-python
langchain
langchain-openai
langchain-community
duckduckgo-search
